--- a/pytest.ini
+++ b/pytest.ini
@@ -1,2 +1,4 @@
 [pytest]
 norecursedirs = archive
+markers =
+    perf: marks tests as performance tests (deselect with '-m "not perf"')