--- a/tests/perf/__init__.py
+++ b/tests/perf/__init__.py
@@ -1,0 +1,1 @@
+# Performance tests package