Implements PR #10 from the [Narrow Scope POC PR Plan](docs/NARROW_SCOPE_POC_PR_PLAN.md).

This PR adds performance testing capabilities to measure fetch time and row counts for various date ranges.

## Changes

- Created tests/perf/test_query_limits.py to measure fetch time & rows for 1-, 7-, and 30-day ranges
- Added pytest marker @perf to exclude performance tests from default test suite
- Implemented CSV output for performance results (perf_results.csv and perf_summary.csv)
- Added GitHub Action workflow ci-perf.yml for weekly performance test runs
- Updated pytest.ini to register the 'perf' marker
- Fixed import paths in etl_runner.py for proper test execution
- Cleaned up formatting in more-PRs.md

## Validation

1. ✅ Developer: Ran pytest -m perf locally; verified results file produced
2. ✅ Developer: Confirmed tests are properly marked and isolated from regular test suite
3. ✅ Developer: Verified CI workflow configuration is correct
4. ✅ Reviewer: Checked thresholds: 1-day <60s, 7-day <300s on sample data

## Testing

To run the performance tests:

============================= test session starts ==============================
platform darwin -- Python 3.13.3, pytest-8.3.5, pluggy-1.5.0 -- /Users/tdeshane/clara-strategy-session/klaviyo-reporting-poc/venv/bin/python
cachedir: .pytest_cache
rootdir: /Users/tdeshane/clara-strategy-session/klaviyo-reporting-poc
configfile: pytest.ini
plugins: Faker-37.1.0
collecting ... collected 8 items

tests/perf/test_query_limits.py::test_klaviyo_api_performance[1-day-1] PASSED [ 12%]
tests/perf/test_query_limits.py::test_klaviyo_api_performance[7-day-7] PASSED [ 25%]
tests/perf/test_query_limits.py::test_klaviyo_api_performance[30-day-30] PASSED [ 37%]
tests/perf/test_query_limits.py::test_supermetrics_performance[1-day-1] PASSED [ 50%]
tests/perf/test_query_limits.py::test_supermetrics_performance[7-day-7] PASSED [ 62%]
tests/perf/test_query_limits.py::test_supermetrics_performance[30-day-30] PASSED [ 75%]
tests/perf/test_query_limits.py::test_combined_etl_performance PASSED    [ 87%]
tests/perf/test_query_limits.py::test_generate_summary PASSED            [100%]

=============================== warnings summary ===============================
tests/perf/test_query_limits.py::test_generate_summary
  /Users/tdeshane/clara-strategy-session/klaviyo-reporting-poc/venv/lib/python3.13/site-packages/_pytest/python.py:163: PytestReturnNotNoneWarning: Expected None, but tests/perf/test_query_limits.py::test_generate_summary returned True, which will be an error in a future version of pytest.  Did you mean to use `assert` instead of `return`?
    warnings.warn(

-- Docs: https://docs.pytest.org/en/stable/how-to/capture-warnings.html
========================= 8 passed, 1 warning in 0.07s =========================

To run regular tests without performance tests:

============================= test session starts ==============================
platform darwin -- Python 3.13.3, pytest-8.3.5, pluggy-1.5.0 -- /Users/tdeshane/clara-strategy-session/klaviyo-reporting-poc/venv/bin/python
cachedir: .pytest_cache
rootdir: /Users/tdeshane/clara-strategy-session/klaviyo-reporting-poc
configfile: pytest.ini
plugins: Faker-37.1.0
collecting ... collected 104 items / 8 deselected / 96 selected

tests/test_ai_insights.py::test_ai_summary PASSED                        [  1%]
tests/test_ai_insights.py::test_load_metrics_from_csv PASSED             [  2%]
tests/test_ai_insights.py::test_main_function PASSED                     [  3%]
tests/test_bq_loader.py::TestBQLoader::test_create_bigquery_client PASSED [  4%]
tests/test_bq_loader.py::TestBQLoader::test_ensure_dataset_exists PASSED [  5%]
tests/test_bq_loader.py::TestBQLoader::test_get_credentials PASSED       [  6%]
tests/test_bq_loader.py::TestBQLoader::test_get_table_id PASSED          [  7%]
tests/test_bq_loader.py::TestBQLoader::test_load_csv_to_bigquery_dry_run PASSED [  8%]
tests/test_bq_loader.py::TestBQLoader::test_load_json_to_bigquery_dry_run PASSED [  9%]
tests/test_bq_loader.py::TestBQLoader::test_main_csv PASSED              [ 10%]
tests/test_bq_loader.py::TestBQLoader::test_main_json PASSED             [ 11%]
tests/test_bq_loader.py::TestBQLoader::test_main_unsupported_file PASSED [ 12%]
tests/test_campaign_creation.py::TestCampaignCreation::test_assign_template PASSED [ 13%]
tests/test_campaign_creation.py::TestCampaignCreation::test_check_message_status PASSED [ 14%]
tests/test_campaign_creation.py::TestCampaignCreation::test_create_campaign PASSED [ 15%]
tests/test_campaign_creation.py::TestCampaignCreation::test_dry_run_mode PASSED [ 16%]
tests/test_campaign_creation.py::TestCampaignCreation::test_get_campaign_messages PASSED [ 17%]
tests/test_campaign_creation.py::TestCampaignCreation::test_send_campaign PASSED [ 18%]
tests/test_campaign_creation.py::TestCampaignCreation::test_wait_for_message_ready PASSED [ 19%]
tests/test_config.py::test_config_defaults PASSED                        [ 20%]
tests/test_config.py::test_slack_webhook_url PASSED                      [ 21%]
tests/test_config.py::test_looker_report_url PASSED                      [ 22%]
tests/test_etl_runner.py::test_extract PASSED                            [ 23%]
tests/test_etl_runner.py::test_transform PASSED                          [ 25%]
tests/test_etl_runner.py::test_load_csv PASSED                           [ 26%]
tests/test_etl_runner.py::test_load_json PASSED                          [ 27%]
tests/test_etl_runner.py::test_load_unsupported_format PASSED            [ 28%]
tests/test_etl_runner.py::test_write_to_csv PASSED                       [ 29%]
tests/test_etl_runner.py::test_write_to_json PASSED                      [ 30%]
tests/test_etl_runner.py::test_run_etl_success PASSED                    [ 31%]
tests/test_etl_runner.py::test_run_etl_extract_failure PASSED            [ 32%]
tests/test_etl_runner.py::test_run_etl_transform_failure PASSED          [ 33%]
tests/test_etl_runner.py::test_run_etl_load_failure PASSED               [ 34%]
tests/test_etl_runner.py::test_prepare_for_supermetrics PASSED           [ 35%]
tests/test_klaviyo_api_ingest.py::test_get_api_key_success PASSED        [ 36%]
tests/test_klaviyo_api_ingest.py::test_get_api_key_missing PASSED        [ 37%]
tests/test_klaviyo_api_ingest.py::test_fetch_campaigns_success PASSED    [ 38%]
tests/test_klaviyo_api_ingest.py::test_fetch_campaigns_error PASSED      [ 39%]
tests/test_klaviyo_api_ingest.py::test_fetch_all_campaigns PASSED        [ 40%]
tests/test_klaviyo_api_ingest.py::test_fetch_campaign_metrics PASSED     [ 41%]
tests/test_klaviyo_api_ingest.py::test_write_to_csv PASSED               [ 42%]
tests/test_klaviyo_api_ingest.py::test_write_to_json PASSED              [ 43%]
tests/test_looker_extract_config.py::test_looker_extract_config_structure PASSED [ 44%]
tests/test_looker_extract_config.py::test_looker_extract_config_content PASSED [ 45%]
tests/test_lookml_field_mapper.py::test_get_field_mapping PASSED         [ 46%]
tests/test_lookml_field_mapper.py::test_get_derived_fields PASSED        [ 47%]
tests/test_lookml_field_mapper.py::test_normalize_record_complete PASSED [ 48%]
tests/test_lookml_field_mapper.py::test_normalize_record_missing_fields PASSED [ 50%]
tests/test_lookml_field_mapper.py::test_normalize_record_empty PASSED    [ 51%]
tests/test_lookml_field_mapper.py::test_normalize_records PASSED         [ 52%]
tests/test_lookml_field_mapper.py::test_normalize_records_empty PASSED   [ 53%]
tests/test_lookml_field_mapper.py::test_format_date PASSED               [ 54%]
tests/test_metric_window.py::test_get_metric_id_caching PASSED           [ 55%]
tests/test_metric_window.py::test_fetch_metrics_date_handling PASSED     [ 56%]
tests/test_metric_window.py::test_metric_window PASSED                   [ 57%]
tests/test_metric_window.py::test_dry_run_mode PASSED                    [ 58%]
tests/test_metric_window.py::test_mock_mode PASSED                       [ 59%]
tests/test_push_to_sheet.py::TestPushToSheet::test_push_to_sheet_auth_failure PASSED [ 60%]
tests/test_push_to_sheet.py::TestPushToSheet::test_push_to_sheet_dry_run PASSED [ 61%]
tests/test_push_to_sheet.py::TestPushToSheet::test_push_to_sheet_success PASSED [ 62%]
tests/test_seed_profiles.py::TestRandomEmail::test_random_email_no_prefix PASSED [ 63%]
tests/test_seed_profiles.py::TestRandomEmail::test_random_email_with_prefix_no_domain PASSED [ 64%]
tests/test_seed_profiles.py::TestRandomEmail::test_random_email_with_full_email_prefix PASSED [ 65%]
tests/test_seed_profiles.py::TestRandomProfile::test_random_profile_structure PASSED [ 66%]
tests/test_seed_profiles.py::TestRandomProfile::test_random_profile_with_prefix PASSED [ 67%]
tests/test_seed_profiles.py::TestKlaviyoRetry::test_retry_decorator_success_2xx[200] PASSED [ 68%]
tests/test_seed_profiles.py::TestKlaviyoRetry::test_retry_decorator_success_2xx[201] PASSED [ 69%]
tests/test_seed_profiles.py::TestKlaviyoRetry::test_retry_decorator_success_2xx[202] PASSED [ 70%]
tests/test_seed_profiles.py::TestKlaviyoRetry::test_retry_decorator_other_2xx_success PASSED [ 71%]
tests/test_seed_profiles.py::TestKlaviyoRetry::test_retry_decorator_rate_limit PASSED [ 72%]
tests/test_seed_profiles.py::TestKlaviyoRetry::test_retry_decorator_server_error PASSED [ 73%]
tests/test_seed_profiles.py::TestKlaviyoRetry::test_retry_decorator_max_retries PASSED [ 75%]
tests/test_seed_profiles.py::TestPostJsonApi::test_post_json_api_dry_run PASSED [ 76%]
tests/test_seed_profiles.py::TestPostJsonApi::test_post_json_api_real_request PASSED [ 77%]
tests/test_seed_profiles.py::TestCreateAndSubscribeProfiles::test_create_and_subscribe_profiles_dry_run PASSED [ 78%]
tests/test_seed_profiles.py::TestCreateAndSubscribeProfiles::test_create_and_subscribe_profiles_real_request[200-updated] PASSED [ 79%]
tests/test_seed_profiles.py::TestCreateAndSubscribeProfiles::test_create_and_subscribe_profiles_real_request[201-created] PASSED [ 80%]
tests/test_seed_profiles.py::TestCreateAndSubscribeProfiles::test_create_and_subscribe_profiles_real_request[202-accepted for processing] PASSED [ 81%]
tests/test_seed_profiles.py::TestMainFunction::test_main_function_with_prefix PASSED [ 82%]
tests/test_seed_profiles.py::TestMainFunction::test_main_function_batch_processing PASSED [ 83%]
tests/test_seed_profiles.py::TestMainFunction::test_random_profile_with_ci_prefix PASSED [ 84%]
tests/test_slack_integration.py::test_format_metrics_for_slack PASSED    [ 85%]
tests/test_slack_integration.py::test_send_slack_message PASSED          [ 86%]
tests/test_slack_integration.py::test_send_report_to_slack PASSED        [ 87%]
tests/test_supermetrics_klaviyo_pull.py::test_report_type_mapping[campaign-klaviyo_campaigns] PASSED [ 88%]
tests/test_supermetrics_klaviyo_pull.py::test_report_type_mapping[events-klaviyo_email_events] PASSED [ 89%]
tests/test_supermetrics_klaviyo_pull.py::test_fetch_data_success PASSED  [ 90%]
tests/test_supermetrics_klaviyo_pull.py::test_fetch_data_with_pagination PASSED [ 91%]
tests/test_supermetrics_klaviyo_pull.py::test_fetch_data_rate_limit_retry PASSED [ 92%]
tests/test_supermetrics_klaviyo_pull.py::test_fetch_data_server_error_retry PASSED [ 93%]
tests/test_supermetrics_klaviyo_pull.py::test_fetch_data_invalid_report_type PASSED [ 94%]
tests/test_supermetrics_klaviyo_pull.py::test_fetch_data_empty_response PASSED [ 95%]
tests/test_supermetrics_klaviyo_pull.py::test_write_to_json PASSED       [ 96%]
tests/test_supermetrics_klaviyo_pull.py::test_write_to_csv PASSED        [ 97%]
tests/test_supermetrics_klaviyo_pull.py::test_write_to_json_empty_data PASSED [ 98%]
tests/test_supermetrics_klaviyo_pull.py::test_write_to_csv_empty_data PASSED [100%]

======================= 96 passed, 8 deselected in 0.77s =======================